+++ lib/arachni/rpc/server/active_options.rb
@@ -18,7 +18,8 @@ class Server
 class ActiveOptions
 
     def initialize( framework )
-        @options = framework.options
+        @framework = framework
+        @options   = framework.options
 
         (@options.public_methods( false ) - public_methods( false ) ).each do |m|
             self.class.class_eval do
@@ -32,7 +33,23 @@ class ActiveOptions
     # @see Arachni::Options#set
     def set( options )
         @options.set( options )
-        HTTP::Client.reset false
+
+        if @framework.running?
+
+            HTTP::Client.reset_options
+
+            # Scope may have been updated.
+            @framework.sitemap.reject! { |k, v| Utilities.skip_path? k }
+
+            @options.scope.extend_paths.each do |url|
+                @framework.push_to_url_queue( url )
+            end
+
+        # Only mess with HTTP state if this is the pre-run config.
+        else
+            HTTP::Client.reset false
+        end
+
         true
     end