+++ spec/arachni/rpc/server/active_options_spec.rb
@@ -6,10 +6,18 @@ require Arachni::Options.paths.lib + 'rpc/server/instance'
 describe Arachni::RPC::Server::ActiveOptions do
     before( :all ) do
         @utils = Arachni::Utilities
-        @instance = instance_spawn
     end
 
-    before( :each ) { @instance.service.clear_cookies }
+    before :each do
+        @instance = instance_spawn
+        @instance.service.clear_cookies
+    end
+    after :each do
+        if @instance
+            @instance.service.shutdown
+            @instance = nil
+        end
+    end
 
     describe '#set' do
         it 'sets options by hash' do
@@ -48,6 +56,56 @@ describe Arachni::RPC::Server::ActiveOptions do
                 Arachni::Cookie.new( url: opts['url'], inputs: { 'name3'  => 'value3' } ),
                 Arachni::Cookie.new( url: opts['url'], inputs: { 'name4'  => 'value4' } )
             ])
+        end
+
+        context 'when the scan is in progress' do
+            before do
+                @url = web_server_url_for( :framework_multi )
+                @instance.service.scan( url: @url )
+            end
+
+            it 'resets HTTP options' do
+                opts = {
+                    'http' => {
+                        'request_concurrency' => 1111
+                    }
+                }
+                @instance.options.set( opts )
+
+                expect(@instance.service.progress[:statistics][:http][:max_concurrency]).to eq 1111
+            end
+
+            it 'removes out of scope sitemap entries' do
+                sleep 1 while @instance.service.sitemap.size < 10
+
+                opts = {
+                    'scope' => {
+                        'exclude_path_patterns' => [
+                            '.'
+                        ]
+                    }
+                }
+                @instance.options.set( opts )
+
+                expect(@instance.service.sitemap).to be_empty
+            end
+
+            it 'pushes new extend paths' do
+                opts = {
+                    'scope' => {
+                        'extend_paths' => [
+                            '/my/path/1',
+                            '/my/path/2'
+                        ]
+                    }
+                }
+                @instance.options.set( opts )
+
+                sleep 1 while @instance.service.busy?
+
+                expect(@instance.service.sitemap).to include "#{@url}/my/path/1"
+                expect(@instance.service.sitemap).to include "#{@url}/my/path/2"
+            end
         end
     end
 end