+++ CHANGELOG.md
@@ -16,6 +16,10 @@
 - `Support`
     - `Signature` -- Optimized signature tokenization, deduplication and compression
         to be less resource intensive when processing large data sets.
+- `RPC`
+    - `Server`
+        - `ActiveOptions#set` -- Allow options to be set during runtime and ajust
+            the scan scope accordingly.
 - `Element`
     - `Capabilities`
         - `Analyzable`